updated auth components and updated signup test
